% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createGrid}
\alias{createGrid}
\title{Create grid}
\usage{
createGrid(gdbPath, lyrName, xdistGrid, ydistGrid, displacementXY = c(0, 0))
}
\arguments{
\item{gdbPath}{The file path (dsn) to the file containing the polygons.}

\item{lyrName}{The layer name for the polygon layer.}

\item{xdistGrid}{Distance in meters between the points in east/west direction.}

\item{ydistGrid}{Distanve in meters between the points in north/south direction.}

\item{displacementXY}{A numeric vector (\code{c(x, y)}), for the displacement of the points.
The origin is at the south west corner of the bounding box of each polygon.}
}
\value{
A simple feature data frame with with a grid layser of points inside the polygons.
}
\description{
Create grid
}
\examples{
grid <- createGrid(gridr_example("best.gpkg"), "best", 20, 20, c(-5, -2))
}
