% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSubsample}
\alias{getSubsample}
\title{Title}
\usage{
getSubsample(grid)
}
\arguments{
\item{grid}{A simple feature point dataframe, probably created by (\code{createGrid()})}
}
\value{
The point data set, together with a column with the distance from one
random chosen point in each polygon and a column with a rank number for the distance.
}
\description{
Title
}
\examples{
grid <- createGrid(gridr_example("best.gpkg"), "best", 20, 20, c(-5, -2))
sg <- getSubsample(grid)
}
