\name{grid.callback}
\alias{grid.callback}
\title{ grid.callback}
\description{
  grid.callback checks it jobs are ready. If so, the result is written to the specified variable and the lock is released.
}
\usage{
grid.callback(...)
}
\arguments{
  \item{...}{for internal use, call this function without parameters}
  }
\details{
This function is automatically called, if you press ENTER. So this should not be called manually.
}
\author{ Malte Lohmeyer }
\keyword{ internal }
