\name{grid.consistency}
\alias{grid.consistency}
\title{ grid.consistency }
\description{
  checks if there are errors in the .grid variable, local files without jobs or jobs without local files.
  Deletes old local tmp files.
}
\usage{
grid.consistency()
}
\author{Malte Lohmeyer }
\keyword{ programming }
