\name{grid.FTPDownload}
\alias{grid.FTPDownload}
\title{ grid.FTPDownload}
\description{
  Downloads a file from url and saves it it path
}
\usage{
grid.FTPDownload(url,path)
}
\arguments{
\item{url}{The url of the file to download}
\item{path}{The path where to download the file}
}
\author{ Malte Lohmeyer  }
\keyword{ file }
