\name{grid.GridFTPDownload}
\alias{grid.GridFTPDownload}
\title{ grid.GridFTPDownload}
\description{
  Downloads a file from url and saves it it path with the use of GridFTP. Only works if <ACGTLIBPATH> in the config file points to the globus toolkit libraries
}
\usage{
grid.GridFTPDownload(url,path)
}
\arguments{
\item{url}{The url of the file to download}
\item{path}{The path where to download the file}
}
\author{ Malte Lohmeyer  }
\keyword{ file }
