\name{grid.GridFTPUpload}
\alias{grid.GridFTPUpload}
\title{ grid.GridFTPUpload}
\description{
  Uploads the file path to url with the use of GridFTP.  Only works if <ACGTLIBPATH> in the config file points to the globus toolkit libraries
}
\usage{
grid.GridFTPUpload(url,path)
}
\arguments{
\item{url}{The url where to upload the file}
\item{path}{The path of the file to upload}
}
\author{ Malte Lohmeyer  }
\keyword{ file }
