\name{grid.restartJob}
\alias{grid.restartJob}
\title{ grid.restartJob}
\description{
  grid.restartJob restarts a Job which is submitted to the scheduler 
}
\usage{
grid.restartJob(job)
}
\arguments{
  \item{job}{The local Variable as String where the output of the Job is saved or the Jobnumber seen by grid.printJobs}
  }
\author{ Malte Lohmeyer }
\keyword{ programming }
