\name{grid.stopJob}
\alias{grid.stopJob}
\title{ grid.stopJob}
\description{
  grid.stopJob stops a Job which is submitted to the scheduler 
}
\usage{
grid.stopJob(job)
}
\arguments{
  \item{job}{The local Variable as String where the output of the Job is saved or the Jobnumber seen by grid.printJobs}
  }
\author{ Malte Lohmeyer }
\keyword{ programming }
