% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_data.R
\name{gridr_example}
\alias{gridr_example}
\title{Get path to example data}
\usage{
gridr_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
There are sample data in `inst/extdata`
directory. This function make them easy to access
Code reused from tidyverse/readr
}
\examples{
gridr_example()
gridr_example("best.gpkg")
}
